refactored categories array into util file
